Added modesRegisterFromDir Function to the existing ModeRegistery file
