+++ src/server.ts
@@ -1,3 +1,4 @@
+import * as path from 'path';
 import './envloader';
 import app from './app';
 import { PORT } from './config/app.config';
@@ -6,19 +7,11 @@ import ModeRegistry from './helpers/modeRegistry';
 import redisClient from './helpers/redisClient';
 import logger from './logger';
 
-import ManualMode from './modes/manual.mode/manual';
-import * as path from 'path';
-import { modesRegisterFromDir } from './utility';
-
 // Initialize the redis client
 redisClient.initialize();
 
-// const manualMode = new ManualMode();
-
 // Register the modes
-// ModeRegistry.registerModes([manualMode]);
-const newModesDir = path.join(__dirname, 'modes');
-modesRegisterFromDir(newModesDir);
+ModeRegistry.modesRegisterFromDir(path.join(__dirname, 'modes'));
 
 app.listen(PORT, () => {
     logger.info(EXPRESS_CONTEXT, `Server running on port ${PORT}`);
